build: Fix browser crashes on daily regression tests
Split teacher_sort_work_view_spec.js into multiple files to prevent browser crashes during daily regression tests.

Each test case is now in its own file.
