+++ cypress/e2e/functional/teacher_tests/teacher_sort_work_view_1_spec.js
@@ -1,0 +1,93 @@
+import TeacherDashboard from "../../../support/elements/common/TeacherDashboard";
+import SortedWork from "../../../support/elements/common/SortedWork";
+import ResourcesPanel from "../../../support/elements/common/ResourcesPanel";
+import Canvas from '../../../support/elements/common/Canvas';
+import ClueHeader from '../../../support/elements/common/cHeader';
+import ChatPanel from "../../../support/elements/common/ChatPanel";
+
+let sortWork = new SortedWork;
+let resourcesPanel = new ResourcesPanel;
+let dashboard = new TeacherDashboard;
+let header = new ClueHeader;
+let chatPanel = new ChatPanel;
+
+const canvas = new Canvas;
+const title = "1.1 Unit Toolbar Configuration";
+const copyTitle = "Personal Workspace";
+const queryParams1 = `${Cypress.config("clueTestqaConfigSubtabsUnitTeacher6")}`;
+
+function beforeTest(params) {
+  cy.visit(params);
+  cy.waitForLoad();
+  dashboard.switchView("Workspace & Resources");
+  cy.wait(2000);
+  cy.openTopTab('sort-work');
+  cy.wait(1000);
+}
+
+//TODO: For QA (1/24)
+// Write a test that confirms correct behavior for "Sort by Tools"
+// • Create a network URL (or clear all documents from existing one from the previous test) that has no documents in Sort Work view (doesn't matter which filter we sort by)
+// • Mock a student (in the same class with a teacher) - have them join the network(when they join the network a problem document is automatically created)
+//   ↳ Next have the student place one tool on the document, lets say "Text"
+//   ↳ As a teacher visit the Sort work view and select the "Sort by Tools" filter, verify that we should see that exact document under the "Text" section label.
+//   ↳ Have the student remove the the Text tool on the document.
+//   ↳ As a teacher again go back to the "Sort by Tools" filter, verify that we see the document under the "No Tools" section label - that is because the student removed the text tool.
+
+describe('SortWorkView Tests', () => {
+  it('should open SortWorkView tab and interact with it', () => {
+    beforeTest(queryParams1);
+    cy.log('verify clicking the sort menu');
+    sortWork.getPrimarySortByMenu().click(); // Open the sort menu
+    cy.wait(500);
+    sortWork.getPrimarySortByNameOption().click(); //Select 'Name' sort type
+    cy.wait(500);
+    sortWork.getPrimarySortByMenu().click(); // Open the sort menu again
+    cy.wait(500);
+    sortWork.getPrimarySortByGroupOption().click(); // Select 'Group' sort type
+    cy.wait(500);
+
+    cy.log('verify opening and closing a document from the sort work view');
+    cy.get('.section-header-arrow').click({multiple: true}); // Open the sections
+    sortWork.getSortWorkItem().eq(1).click(); // Open the first document in the list
+    resourcesPanel.getEditableDocumentContent().should('be.visible');
+
+    cy.log('verify document scroller is visible, populated, and functions');
+    let prevFocusDocKey = "";
+    let selectedDocIndex = 0;
+    resourcesPanel.getEditableDocumentContent().invoke('attr', 'data-focus-document').then((focusDocKey) => {
+      prevFocusDocKey = focusDocKey;
+    });
+    resourcesPanel.getDocumentScroller().should('be.visible').and($el => {
+      expect($el.find('[data-testid="document-thumbnail"]')).to.have.length.greaterThan(1);
+      expect($el.find('[data-testid="document-thumbnail"].selected')).to.have.length(1);
+      selectedDocIndex = $el.find('[data-testid="document-thumbnail"]')
+                         .index($el.find('[data-testid="document-thumbnail"].selected'));
+    });
+    resourcesPanel.getDocumentScrollerLeftBtn().should('not.exist');
+    cy.get('[data-testid="document-thumbnail"]').first().should('be.visible');
+    resourcesPanel.getDocumentScrollerRightBtn().should('exist').click();
+    cy.get('[data-testid="document-thumbnail"]').first().should('not.be.visible');
+    resourcesPanel.getDocumentScrollerLeftBtn().should('exist').click();
+    cy.get('[data-testid="document-thumbnail"]').first().should('be.visible');
+    cy.get('[data-testid="document-thumbnail"]').eq(selectedDocIndex + 1).click();
+    resourcesPanel.getEditableDocumentContent().invoke('attr', 'data-focus-document')
+                                               .should('not.eq', prevFocusDocKey).then((focusDocKey) => {
+                                                 prevFocusDocKey = focusDocKey;
+                                               });
+
+    cy.log('verify document scroller is collapsible, and that switch document buttons appear when it is collapsed');
+    resourcesPanel.getDocumentSwitchBtnPrev().should('not.exist');
+    resourcesPanel.getDocumentSwitchBtnNext().should('not.exist');
+    resourcesPanel.getDocumentScrollerToggle().should('exist').click();
+    resourcesPanel.getDocumentScroller().should('not.exist');
+    resourcesPanel.getDocumentSwitchBtnPrev().should('exist').and('not.have.class', 'disabled').click();
+    resourcesPanel.getDocumentSwitchBtnPrev().should('have.class', 'disabled');
+    resourcesPanel.getEditableDocumentContent().invoke('attr', 'data-focus-document')
+                                               .should('not.eq', prevFocusDocKey);
+    resourcesPanel.getDocumentSwitchBtnNext().should('exist').and('not.have.class', 'disabled');
+
+    resourcesPanel.getDocumentCloseButton().click();
+    sortWork.getSortWorkItem().should('be.visible'); // Verify the document is closed
+  });
+});
